+++ src/main/java/br/com/fiap/esgsmart/repository/ReciclagemRepository.java
@@ -6,13 +6,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface ReciclagemRepository extends JpaRepository<Reciclagem, Integer> {
+public interface ReciclagemRepository extends JpaRepository<Reciclagem, Long> {
 
     List<Reciclagem> findByTipoMaterial(String tipoMaterial);
 
     List<Reciclagem> findByQuantidadeReciclada(double quantidade);
 
     List<Reciclagem> findByDataReciclagem(LocalDate data);
-
 
 }
